app/requirements.txt
Flask==1.1.2
google-cloud-bigquery==2.5.0
pandas==1.1.4
pyarrow==2.0.0